Move pool converter out of curve.rs
